feat: Enable Nuxt Pages and update layout structure
This commit enables the Nuxt Pages feature to allow for file-based routing.
The default app layout is updated to include `<NuxtLayout>` and `<NuxtPage>` components,
which enables the use of custom layouts and page rendering.
Also, the default favicon.ico was removed.
Finally, the i18n configuration was updated to disable `optimizeTranslationDirective`.

This change is required to support a multi-page application structure and improve
the flexibility of page layouts.